refactor: renamed BusTicket packages into bus_tickets
